Add "real run" awareness to tag push operation
This adds the `is_real_run` logic we use in our `grapl-rc` Buildkite
plugin to this plugin. If we're doing a "test run" of any pipeline
that uses this plugin, we don't want to inadvertently push tags to
Github; that should only happen for "real" pipeline runs (here, "real"
is currently defined as any build not triggered via the UI).

+++ lib/record.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
+# shellcheck source-path=SCRIPTDIR
+source "$(dirname "${BASH_SOURCE[0]}")/util.sh"
+
 # Assuming our pipeline names are of the form of:
 #
 #     REPOSITORY_NAME/PIPELINE
@@ -36,6 +39,11 @@ tag_last_success() {
 
     echo "--- :github: Re-tagging '${tag}'"
     git tag "${tag}" --force
-    echo "--- :github: Pushing new value for '${tag}' tag"
-    git push origin "${tag}" --force --verbose
+
+    if is_real_run; then
+        echo "--- :github: Pushing new value for '${tag}' tag"
+        git push origin "${tag}" --force --verbose
+    else
+        echo -e "--- :no_good: Would have pushed '${tag}' tag to Github"
+    fi
 }
